+++ .babelrc
@@ -1,5 +1,5 @@
 {
   "presets": ["env", "react"],
-  "plugins": ["transform-class-properties"]
+  "plugins": ["transform-class-properties", "transform-object-rest-spread"]
 
 }
